move Directory interface to separate file
